Crate custom_store_subscriber

This example demonstrates how to use ChunkStoreSubscribers and ChunkStoreEvents to implement both custom secondary indices and trigger systems.

Usage:

# Start the Rerun Viewer with our custom view in a terminal:
$ cargo r -p custom_store_subscriber

# Log any kind of data from another terminal:
$ cargo r -p objectron -- --connect

Structs

ComponentsPerRecording 🔒

A ChunkStoreSubscriber that maintains a secondary index that keeps count of the number of occurrences of each component in each rerun::ChunkStore.

Orchestrator 🔒

A meta ChunkStoreSubscriber that distributes work to our other views.

TimeRangesPerEntity 🔒

A ChunkStoreSubscriber that maintains a secondary index of the time ranges covered by each entity, on every timeline, across all recordings (i.e. rerun::ChunkStores).

Functions

main 🔒